mobile: debug log bottomsheet copy+style
release: 1.4.1 build 85
