+++ apps/daimo-mobile/src/common/useSendDebugLog.ts
@@ -2,54 +2,14 @@ import { assertNotNull } from "@daimo/common";
 import { daimoChainFromId } from "@daimo/contract";
 import * as ExpoEnclave from "@daimo/expo-enclave";
 import * as FileSystem from "expo-file-system";
-import { useState, useEffect } from "react";
-import { Platform, ShareContent, Share } from "react-native";
+import { useEffect, useState } from "react";
+import { Platform, Share, ShareContent } from "react-native";
 
-import { getHardwareSec } from "./logic/enclave";
-import { env } from "./logic/env";
-import { Account, getAccountManager, serializeAccount } from "./model/account";
-import { amountSeparator } from "./view/shared/Amount";
-
-const n = 400;
-const logs: string[] = [];
-
-// Hooks console.log, saves a rolling log of the last n lines.
-export function initDebugLog() {
-  console.log = createLogFunc("log", console.log);
-  console.warn = createLogFunc("WRN", console.warn);
-  console.error = createLogFunc("ERR", console.error);
-}
-
-function createLogFunc(type: string, oldLog: (...args: any[]) => void) {
-  return (...args: any[]) => {
-    // Print to console in local development.
-    oldLog(...args);
-
-    // Save to rolling buffer.
-    const timestamp = new Date().toISOString();
-    const parts = args.map((a) => {
-      if (typeof a === "string") return a;
-      if (a instanceof Error) return a.stack || a.toString();
-      try {
-        return JSON.stringify(a);
-      } catch {
-        return "" + a;
-      }
-    });
-    let line = [timestamp, type, ...parts].join(" ");
-    if (line.length > 5000) line = line.slice(0, 5000) + "...";
-    logs.push(line);
-
-    // Don't let the buffer get too long
-    if (logs.length > n) logs.shift();
-  };
-}
-
-export function getDebugLog(headerLines: string[]) {
-  const now = new Date().toISOString();
-  const log = logs.join("\n") + `\n${now} - debug log captured`;
-  return [`# Daimo Debug Log`, ...headerLines, log].join("\n\n");
-}
+import { getDebugLog } from "./debugLog";
+import { getHardwareSec } from "../logic/enclave";
+import { env } from "../logic/env";
+import { Account, getAccountManager, serializeAccount } from "../model/account";
+import { amountSeparator } from "../view/shared/Amount";
 
 /**
  * Returns a function to send a complete debug log + env summary.
@@ -107,7 +67,6 @@ export function useSendDebugLog(
 
   return [sendDebugLog, envKV];
 }
-
 function getKeySecDescription(
   hardwareSecurityLevel: ExpoEnclave.HardwareSecurityLevel
 ) {
